requirements.txt
flask
psycopg2-binary
gunicorn
requests
flask-limiter
flask-cors